Fix viewer count and time elapsed
